requirements.txt
altair==2.1.0
arrow==0.12.1
asn1crypto==0.24.0
awscli==1.14.58
beautifulsoup4==4.6.0
binaryornot==0.4.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
colorama==0.3.7
cryptography==2.2.2
cycler==0.10.0
decorator==4.3.0
descartes==1.1.0
docutils==0.14
et-xmlfile==1.0.1
feedparser==5.2.1
Fiona==1.7.11.post1
future==0.16.0
geopandas==0.3.0
google-api-python-client==1.6.5
haversine==0.4.5
holidays==0.9.5
html5lib==1.0.1
httplib2==0.10.3
idna==2.6
ipython-genutils==0.2.0
jdcal==1.4
Jinja2==2.10
jinja2-time==0.2.0
jmespath==0.9.3
jsonschema==2.6.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.2.0
munch==2.2.0
nbformat==4.4.0
numpy==1.14.1
oauth2client==4.1.2
openpyxl==2.5.3
pandas==0.22.0
poyo==0.4.1
pprint==0.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDrive==1.3.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyproj==1.9.5.1
pypyodbc==1.3.4
python-dateutil==2.6.1
python-dotenv==0.8.2
python-forecastio==1.3.5
pytz==2018.3
PyYAML==3.12
ratelim==0.1.6
requests==2.18.4
requests-toolbelt==0.8.0
responses==0.8.1
rsa==3.4.2
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
selenium==3.8.0
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
webencodings==0.5.1
whichcraft==0.4.1
xlrd==1.1.0
yellowbrick==0.7
